Specify COLUMNS parameter in "ahoy drush" to prevent output wrapping
This is in order to address this issue:

![image](https://user-images.githubusercontent.com/2423362/173475570-2ac6fe2c-8095-4c20-98ce-052219f9a257.png)

As you can see from the above, running `ahoy drush uinf` directly returns a wrapped username/email, which is hard to copy/paste or parse via grep. Whereas first ssh'ing via `ahoy cli` and then running `drush uinf` has the correctly-formatted, expected output.

This is not an issue with ahoy, rather than with docker-compose, and is can be fixed is you specify the `COLUMNS` parameter. I've set it to 120, which basically increases it from the default 80 and seems plenty (should fix the issue in most cases):

![image](https://user-images.githubusercontent.com/2423362/173477707-833f29cc-4b73-40fa-9e94-757907d3c043.png)